- Laravel upgraded to v8 - Migrations updated - Model relationships - Home page base data display - Minor changes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->date('birth_date');
             $table->string('avatar');
-            $table->unsignedBigInteger('country');
+            $table->integer('country');
             $table->integer('points');
             $table->boolean('is_admin');
             $table->boolean('is_banned');
